extract scrollbar to separate component
