+++ src/components/index.ts
@@ -5,5 +5,6 @@ import {PrePlaybackPlayOverlayWrapper} from './pre-playback-play-overlay-wrapper
 import {RelatedList} from './related-list/related-list';
 import {CloseButton} from './close-button/close-button';
 import {RelatedContext} from './related-context/related-context';
+import {Scrollable} from './scrollable/scrollable';
 
-export {Next, PrePlaybackPlayOverlayWrapper, RelatedOverlay, ListToggleButton, RelatedList, CloseButton, RelatedContext};
+export {Next, PrePlaybackPlayOverlayWrapper, RelatedOverlay, ListToggleButton, RelatedList, CloseButton, RelatedContext, Scrollable};
