+++ src/components/related-list/related-list.scss
@@ -1,24 +1,3 @@
-@mixin related-scrollbar {
-    &::-webkit-scrollbar {
-        height: 4px;
-        width: 4px;
-    }
-    &::-webkit-scrollbar-track {
-        background: rgba(33, 33, 33, 0.9);
-        visibility: hidden;
-    }
-    &::-webkit-scrollbar-thumb {
-        background-color: rgba(255, 255, 255, 0.3);
-        border-radius: 3px;
-        visibility: hidden;
-    }
-
-    &.scrolling::-webkit-scrollbar-track,
-    &.scrolling::-webkit-scrollbar-thumb {
-        visibility: visible;
-    }
-}
-
 .related-list {
     -webkit-backdrop-filter: blur(8px);
             backdrop-filter: blur(8px);
@@ -45,9 +24,7 @@
         }
     }
 
-    .content {
-        @include related-scrollbar();
-        
+    .content {        
         border: 1px solid red;
         color: red;
         display: flex;
